Final changes, ready for the hive
Got rid of json lines, we do not need them. Set up so there is 1 measurement taken every 10 minutes, and uploaded every 2 hours. The first 100 data points will be taken and tagged as "t-data" = test data, and the rest "u-data" = use data. We should be able to filter by the tag on blues notehub.